requirements.txt
asgiref==3.8.1
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
folium==0.19.7
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
polyline==2.0.2
python-dotenv==1.1.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
xyzservices==2025.4.0
